+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Add new config variable `flowman.default.relation.output.columnMismatchPolicy` (default is `ADD_REMOVE_COLUMNS`)
 * Add new config variable `flowman.default.relation.output.typeMismatchPolicy` (default is `CAST_ALWAYS`)
 * Improve handling of `_SUCCESS` files for detecting (non-)dirty directories
+* Implement new `merge`target
+* Implement merge operation for Delta relations
 
 
 # Version 0.19.0 - 2021-12-13